fix: moved project specific features to project folder
