+++ src/app/dashboard/projects/page.tsx
@@ -1,5 +1,5 @@
 import { DashboardPageHeader } from '@/components/dashboard/page-header'
-import { ProjectsPageClient } from '@/components/dashboard/projects-page-client'
+import { ProjectsPageClient } from '@/components/dashboard/project/projects-page-client'
 
 export default async function ProjectsPage() {
   return (
