+++ src/components/dashboard/project/project-detail-client.tsx
